Revert "Moved Avatar (PersonaCoin) to framework v2. (#1161)"
This reverts commit 6f1258d182196549bdd747a752ab8c28e5b60e2a.
