+++ apps/fluent-tester/src/FluentTester/TestComponents/Avatar/consts.ts
@@ -1,3 +1,2 @@
 export const HOMEPAGE_AVATAR_BUTTON = 'Homepage_Avatar_Button';
 export const AVATAR_TESTPAGE = 'Avatar_TestPage';
-export const AVATAR_WIN32_TESTPAGE = 'Avatar_Win32_TestPage';
