+++ apps/fluent-tester/src/FluentTester/TestComponents/Avatar/index.ts
@@ -1,3 +1,2 @@
 export * from './AvatarTest';
-export * from './AvatarTestWin32';
 export * from './consts';
