+++ apps/fluent-tester/src/FluentTester/testPages.win32.ts
@@ -1,4 +1,3 @@
-import { AvatarTestWin32, HOMEPAGE_AVATAR_BUTTON } from './TestComponents/Avatar';
 import { ButtonTest, HOMEPAGE_BUTTON_BUTTON } from './TestComponents/Button';
 import { ExperimentalButtonTest, HOMEPAGE_BUTTON_BUTTONEXPERIMENTAL } from './TestComponents/ButtonExperimental';
 import { CalloutTest, HOMEPAGE_CALLOUT_BUTTON } from './TestComponents/Callout';
@@ -27,11 +26,6 @@ import { HOMEPAGE_CHECKBOX_CHECKBOXEXPERIMENTAL, ExperimentalCheckboxTest } from
 import { ExperimentalMenuButtonTest, HOMEPAGE_MENU_BUTTON_EXPERIMENTAL } from './TestComponents/MenuButtonExperimental';
 
 export const tests: TestDescription[] = [
-  {
-    name: 'Avatar Test',
-    component: AvatarTestWin32,
-    testPage: HOMEPAGE_AVATAR_BUTTON,
-  },
   {
     name: 'Button Test',
     component: ButtonTest,
