--- change/@fluentui-react-native-experimental-avatar-8531a6ad-48b7-4984-bc66-2ed9e31f3a79.json
@@ -1,7 +1,0 @@
-{
-  "type": "patch",
-  "comment": "Moved Avatar (PersonaCoin) to framework v2.",
-  "packageName": "@fluentui-react-native/experimental-avatar",
-  "email": "v.kozlova13@gmail.com",
-  "dependentChangeType": "patch"
-}
