--- change/@fluentui-react-native-tester-2e77557e-87ec-4dae-9c33-31c9e06d7fce.json
@@ -1,7 +1,0 @@
-{
-  "type": "patch",
-  "comment": "Moved Avatar (PersonaCoin) to framework v2.",
-  "packageName": "@fluentui-react-native/tester",
-  "email": "v.kozlova13@gmail.com",
-  "dependentChangeType": "patch"
-}
