+++ packages/experimental/Avatar/package.json
@@ -24,22 +24,18 @@
     "prettier-fix": "fluentui-scripts prettier --fix true"
   },
   "dependencies": {
-    "@fluentui-react-native/adapters": "^0.7.5",
     "@fluentui-react-native/component-cache": "^1.2.3",
-    "@fluentui-react-native/framework": "0.6.6",
-    "@fluentui-react-native/theme-tokens": "^0.9.2",
-    "@fluentui-react-native/tokens": "^0.10.3",
-    "@fluentui-react-native/use-styling": "^0.7.2"
+    "@fluentui-react-native/framework": "0.6.6"
   },
   "devDependencies": {
     "@fluentui-react-native/eslint-config-rules": "^0.1.1",
     "@types/react-native": "^0.63.0",
     "@uifabricshared/build-native": "^0.1.1",
-    "react": "16.13.1",
-    "react-native": "^0.63.4"
+    "react-native": "^0.63.4",
+    "react": "16.13.1"
   },
   "peerDependencies": {
-    "react": "16.13.1",
-    "react-native": ">=0.63.4"
+    "react-native": ">=0.63.4",
+    "react": "16.13.1"
   }
 }
