+++ packages/experimental/Avatar/src/Avatar.tsx
@@ -1,44 +1,134 @@
 /** @jsx withSlots */
-import { Image, View, Text } from 'react-native';
-import { AvatarProps, AvatarType, avatarName } from './Avatar.types';
-import { stylingSettings } from './Avatar.styling';
-import { compose, UseSlots, mergeProps, withSlots } from '@fluentui-react-native/framework';
-import { useAvatar } from './useAvatar';
+import { compose, UseSlots, buildProps, mergeProps, withSlots } from '@fluentui-react-native/framework';
+import { ImageURISource, NativeModules, ViewProps, ColorValue } from 'react-native';
+import { ensureNativeComponent } from '@fluentui-react-native/component-cache';
+
+const avatarName = 'Avatar';
+
+const NativeAvatarView = ensureNativeComponent('FRNAvatarView');
+
+export type Size = 'xSmall' | 'small' | 'medium' | 'large' | 'xLarge' | 'xxLarge';
+
+export type AvatarStyle = 'default' | 'accent' | 'group' | 'outlined' | 'outlinedPrimary' | 'overflow';
+
+export type Presence = 'none' | 'available' | 'away' | 'blocked' | 'doNotDisturb' | 'offline' | 'unknown';
+
+interface ExportedConstants {
+  sizes: { [key in Size]: number };
+}
+
+const ExportedNativeConstants: ExportedConstants = NativeModules.FRNAvatarViewManager;
+
+export type AvatarTokens = {
+  /**
+   * Supported Avatar sizes
+   */
+  size?: Size;
+
+  /**
+   * Supported Avatar styles
+   */
+  avatarStyle?: AvatarStyle;
+};
+
+export type AvatarProps = AvatarTokens & {
+  /**
+   * The image to be displayed
+   */
+  imageSource?: ImageURISource;
+
+  /**
+   * The primary text to create initials with (e.g. a name)
+   */
+  primaryText?: string;
+
+  /**
+   * The secondary text to create initials with if primary text is not provided (e.g. an email address)
+   */
+  secondaryText?: string;
+
+  /**
+   * A custom color for the ring around the Avatar
+   */
+  ringColor?: ColorValue;
+
+  /**
+   * A custom color for the avatar initials when no image is shown
+   */
+  foregroundColor?: ColorValue;
+
+  /**
+   * A custom color for the avatar background when no image is shown
+   */
+  backgroundColor?: ColorValue;
+
+  /**
+   * The avatar view's presence state.
+   */
+  presence?: Presence;
+
+  /**
+   * A boolean inidicating whether the ring is visible
+   */
+  isRingVisible?: boolean;
+
+  /**
+   * A boolean indicating whether the Inner Ring Gap is transparent
+   */
+  isTransparent?: boolean;
+
+  /**
+   * A boolean indicating out of office status
+   */
+  isOutOfOffice?: boolean;
+
+  /**
+   * A boolean indicating whether the gap between the avatar and ring is visible.
+   */
+  hasRingInnerGap?: boolean;
+
+  /**
+   * An image to be shown as the backdrop of the ring, rather then a solid color.
+   * Takes precendence over `ringColor`
+   */
+  customBorderImageSource?: ImageURISource;
+};
+
+const tokensThatAreAlsoProps: (keyof AvatarTokens)[] = ['size', 'avatarStyle'];
+
+export type NativeAvatarViewProps = ViewProps & AvatarProps;
+interface AvatarType {
+  props: AvatarProps;
+  slotProps: { root: NativeAvatarViewProps };
+  tokens: AvatarTokens;
+}
 
 export const Avatar = compose<AvatarType>({
   displayName: avatarName,
-  ...stylingSettings,
-  slots: {
-    root: View,
-    photo: Image,
-    initials: Text,
-    initialsBackground: View,
-    icon: Image,
-    ring: View,
-    glow: View,
+  tokens: [
+    {
+      size: 'small',
+      avatarStyle: 'default',
+    },
+    avatarName,
+  ],
+  tokensThatAreAlsoProps,
+  slots: { root: NativeAvatarView },
+  slotProps: {
+    root: buildProps<NativeAvatarViewProps, AvatarTokens>(
+      (tokens) => ({
+        size: tokens.size,
+        style: {
+          height: ExportedNativeConstants.sizes[tokens.size],
+          width: ExportedNativeConstants.sizes[tokens.size],
+        },
+      }),
+      ['size'],
+    ),
   },
-  render: (userProps: AvatarProps, useSlots: UseSlots<AvatarType>) => {
-    const Avatar = useAvatar(userProps);
-    const Slots = useSlots(userProps, (layer) => Avatar.state[layer] || userProps[layer]);
+  render: (props: AvatarProps, useSlots: UseSlots<AvatarType>) => {
+    const Root = useSlots(props).root;
 
-    return (final: AvatarProps) => {
-      const { children, accessibilityLabel, ...mergedProps } = mergeProps(Avatar.props, final);
-      const { personaPhotoSource, iconSource, showRing, transparentRing } = Avatar.state;
-
-      return (
-        <Slots.root {...mergedProps}>
-          {personaPhotoSource ? (
-            <Slots.photo accessibilityLabel={accessibilityLabel} source={personaPhotoSource} />
-          ) : (
-            <Slots.initialsBackground>
-              <Slots.initials>{children}</Slots.initials>
-            </Slots.initialsBackground>
-          )}
-          {showRing && !transparentRing && <Slots.ring />}
-          {showRing && <Slots.glow />}
-          {!!iconSource && !!iconSource.uri && <Slots.icon source={iconSource} />}
-        </Slots.root>
-      );
-    };
+    return (rest: AvatarProps) => <Root {...mergeProps(props, rest)} />;
   },
 });
