+++ packages/experimental/Avatar/src/index.ts
@@ -1,2 +1,1 @@
 export * from './Avatar';
-export * from './Avatar.types';
